smileyCTF2025/pyproject.toml
[project]
name = "smiley"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pwn>=1.0",
    "pycryptodome>=3.23.0",
    "z3>=0.2.0",
    "z3-solver>=4.15.1.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.16.1",
    "ruff>=0.12.0",
]
